backend/requirements.txt
# FastAPI
fastapi

# Uvicorn
uvicorn

# llama-index dependencies
llama-index-core
llama-index-vector-stores-weaviate
llama-index-embeddings-huggingface
llama-index-readers-file
llama-index-llms-openrouter

# OpenAI
openai

# Weaviate client
weaviate-client

# Caching
redis

# URL handling
aiohttp
beautifulsoup4

# Settings
pydantic-settings

# Database
asyncpg
sqlalchemy
alembic

# JWT dependencies
PyJWT
passlib==1.7.4
bcrypt==3.2.2

# Backoff
backoff

# File handling
python-magic
python-multipart

# Environment variables
python-dotenv

# Documents handling
docx2txt

# dotenv
python-dotenv
